Fix CI failures: resolve React hooks rules and ESLint warnings
- Move useAuth hook to separate file to resolve React Fast Refresh warning
- Fix React hooks rules violations in LeadDetail component by moving hooks before conditional returns
- Update all imports to use new useAuth location
- Ensure all linting and build issues are resolved
